requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
frozenlist==1.7.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.28
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.107.3
orjson==3.11.3
packaging==25.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.32.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.25.0
